requirements.txt
requests>=2.31.0
pystray>=0.19.4
Pillow>=10.0.0
winotify>=1.1.0
win10toast>=0.9
plyer>=2.1.0